feat(auth): Refactor LoginRequestDTO and update imports
- Moved `LoginRequestDTO` to a new package under `dtos` to better organize the project structure.
- Updated import paths for `LoginRequestDTO` in `AuthController` and `JwtTokenService` to reflect the new location.
- Deleted the old file from the `controllers.v1` package.
- Added an `ignore` section in `codecov.yml` to exclude specific directories and files from coverage reporting, including models, dtos, exception classes, and security configurations.

This refactor consolidates the DTOs in a dedicated package, improving the maintainability of the codebase.
