+++ server/src/main/kotlin/no/uyqn/server/controllers/v1/AuthController.kt
@@ -6,6 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
+import no.uyqn.server.dtos.LoginRequestDTO
 import no.uyqn.server.dtos.TokenDTO
 import no.uyqn.server.services.JwtTokenService
 import org.springframework.web.bind.annotation.PostMapping
